Make Vigor baseline setup.sh idempotent

+++ baselines/vigor/setup.sh
@@ -7,9 +7,13 @@ git submodule update --init --recursive
 sudo apt-get install -y libnuma-dev
 
 cd dpdk
+# Vigor's patches are not idempotent!
+if [ ! -f .built ]; then
+  for p in ../vigor/setup/dpdk.*.patch; do
+    patch -p1 < "$p"
+  done
 
-for p in ../vigor/setup/dpdk.*.patch; do
-  patch -p1 < "$p"
-done
+  make install -j$(nproc) T=x86_64-native-linuxapp-gcc DESTDIR=.
 
-make install -j$(nproc) T=x86_64-native-linuxapp-gcc DESTDIR=.
+  touch .built
+fi
